f/mmcdaniel-baroreceptors: Set opiods to effect chemoreceptor signal
CNS drug modifier was being applied to output of chemoreceptor function
(total ventilation target) but not to the chemoreceptor signal itself.
The signal kept increasing during fentanyl and morphine scenarios, which
caused improper downstream effects in Central Signal and Efferent
response functions.

A similar issue with the Anesthesia Machine still needs to be resolved.