f/mmcdaniel-baroreceptors: Improve heart rate calculation
The cardiac cycle reset can leave some time unaccounted for if target
cycle period is not a multiple of the time step.  Add this back to
variable tracking beat duration so that heart rate calculation knows to
include it.  Since an error of < 1 time step can introduce perturbations
in heart rate that prevent stabilization, do not update heart period
unless the input is > 1 time step.